+++ src/features/TagAction/ui/TagOptions/DeleteTagItem.tsx
@@ -1,19 +1,10 @@
-import {
-	Button,
-	MenuItem,
-	Modal,
-	ModalContent,
-	ModalFooter,
-	ModalOverlay,
-	Text,
-} from "@chakra-ui/react"
 import { useEvent } from "effector-react"
 import { FC } from "react"
 import { RiDeleteBin6Line } from "react-icons/ri"
 
 import { $tagsApi } from "@/shared/api"
-import { MODAL_CONFIRM_CANCEL, MODAL_CONFIRM_OK } from "@/shared/lib"
-import { useDisclosure } from "@/shared/package/react-hooks"
+import { useConfirm } from "@/shared/package/react-confirm"
+import { MenuItem } from "@/shared/ui"
 
 import { CONFIRM_DELETE_TEXT, DELETE_TAG_ITEM } from "../../lib"
 
@@ -25,31 +16,20 @@ interface DeleteTagItemProps {
 }
 
 export const DeleteTagItem: FC<DeleteTagItemProps> = ({ id, onDelete }) => {
-	const { isOpen, onClose, onOpen } = useDisclosure()
+	const confirm = useConfirm()
 	const deleteTag = useEvent($tagsApi.deleteTag)
 
 	const clickHandle = () => {
-		deleteTag(id)
-		onClose()
-		onDelete && onDelete()
+		confirm(CONFIRM_DELETE_TEXT, () => {
+			deleteTag(id)
+			onDelete && onDelete()
+		})
 	}
 
 	return (
-		<>
-			<MenuItem onClick={onOpen} className={styles.deleteItem}>
-				<RiDeleteBin6Line className={styles.itemIcon} />
-				{DELETE_TAG_ITEM}
-			</MenuItem>
-			<Modal isOpen={isOpen} onClose={onClose}>
-				<ModalOverlay />
-				<ModalContent className={styles.modal}>
-					<Text>{CONFIRM_DELETE_TEXT}</Text>
-					<ModalFooter className={styles.modalFooter}>
-						<Button onClick={onClose}>{MODAL_CONFIRM_CANCEL}</Button>
-						<Button onClick={clickHandle}>{MODAL_CONFIRM_OK}</Button>
-					</ModalFooter>
-				</ModalContent>
-			</Modal>
-		</>
+		<MenuItem onClick={clickHandle} className={styles.deleteItem}>
+			<RiDeleteBin6Line className={styles.itemIcon} />
+			{DELETE_TAG_ITEM}
+		</MenuItem>
 	)
 }
